image_retrieval/requirements.txt
streamlit
PyYAML==6.0.1
pandas
dataclasses
tqdm
nltk
num2words
